Separate demos declarative and programmatic init
