+++ demos/src/declarative.js
@@ -1,0 +1,16 @@
+/*global require*/
+import oPermutive from '../../main.js';
+
+// document.addEventListener('DOMContentLoaded', function() {
+// 	document.dispatchEvent(new CustomEvent('o.DOMContentLoaded'));
+// });
+const oPermConf = {
+	"appInfo": {
+		"appName": "article",
+		"contentId": "5cfae92e-6cc5-11e9-80c7-60ee53e6681d"
+	},
+};
+
+document.cookie = 'FTConsent=behaviouraladsOnsite%3Aon;';
+
+//oPermutive.pAddon({}, oPermConf);
