+++ demos/src/declarative.mustache
@@ -1,0 +1,7 @@
+<div class="o-permutive"
+	data-o-component="o-permutive"
+	data-o-permutive-publicApiKeys-id="{{permutiveApiId}}"
+	data-o-permutive-publicApiKeys-key="{{permutiveApiKey}}"
+	data-o-permutive-adsApi-user="{{adsApiUser}}"
+	data-o-permutive-adsApi-content="{{adsApiContent}}"
+></div>
