--- demos/src/demo.mustache
@@ -1,9 +1,0 @@
-<div>
-	<div class="o-permutive" data-o-component="o-permutive"
-	data-o-permutive-auoto-init="false"
-	data-o-permutive-id="e1c3fd73-dd41-4abd-b80b-4278d52bf7aa"
-	data-o-permutive-key="b2b3b748-e1f6-4bd5-b2f2-26debc8075a"
-	data-o-permutive-page-type="article"
-	data-o-permutive-apiendpoint-user="https://ads-api.ft.com/v1/user"
-	data-o-permutive-apiendpoint-content="https://ads-api.ft.com/v1/content/"></div>
-</div>
