+++ demos/src/programmatic.mustache
@@ -1,0 +1,2 @@
+<div class="o-sweet-nothing">
+</div>
