--- index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-	<title>Getting Started</title>
-	<meta
-		name="o-permutive"
-		class="o-permutive"
-		data-o-component="o-permutive"
-		data-o-permutive-publicApiKeys-id="e1c3fd73-dd41-4abd-b80b-4278d52bf7aa"
-		data-o-permutive-publicApiKeys-key="b2b3b748-e1f6-4bd5-b2f2-26debc8075a"
-		data-o-permutive-pageType="article"
-		data-o-permutive-adsApi-user="https://ads-api.ft.com/v1/user"
-		data-o-permutive-adsApi-content="https://ads-api.ft.com/v1/content/5cfae92e-6cc5-11e9-80c7-60ee53e6681d"
-		data-o-permutive-appInfo-contentId="b2b3b748-e1f6-4bd5-b2f2-26debc8075a"
-	>
-	</meta>
-</head>
-
-<body>
-	<script src="./build/main.js"></script>
-	<script>
-		document.addEventListener('DOMContentLoaded', function () {
-			document.dispatchEvent(new CustomEvent('o.DOMContentLoaded'));
-		});
-	</script>
-</body>
-
-</html>
