+++ origami.json
@@ -19,15 +19,29 @@
 	},
 	"demosDefaults": {
 		"sass": "demos/src/demo.scss",
-		"js": "demos/src/demo.js",
+		"js": "demos/src/declarative.js",
 		"documentClasses": "",
 		"dependencies": ""
 	},
 	"demos": [
 		{
-			"title": "Demo",
-			"name": "Set config declaritively",
-			"template": "demos/src/demo.mustache",
+			"js": "demos/src/declarative.js",
+			"title": "Declarative Demo",
+			"name": "Set config declaratively",
+			"template": "demos/src/declarative.mustache",
+			"description": "",
+			"data": {
+				"permutiveApiId": "e1c3fd73-dd41-4abd-b80b-4278d52bf7aa",
+				"permutiveApiKey": "b2b3b748-e1f6-4bd5-b2f2-26debc8075a",
+				"adsApiUser": "https://ads-api.ft.com/v1/user",
+				"adsApiContent": "https://ads-api.ft.com/v1/content/5cfae92e-6cc5-11e9-80c7-60ee53e6681d"
+			}
+		},
+		{
+			"js": "demos/src/programmatic.js",
+			"title": "Programmatic Demo",
+			"name": "Set config programmatically",
+			"template": "demos/src/programmatic.mustache",
 			"description": ""
 		}
 	]
